+++ banzai_floyds/frames.py
@@ -1,5 +1,4 @@
 from banzai.lco import LCOObservationFrame, LCOCalibrationFrame, LCOFrameFactory
-from banzai.frames import CalibrationFrame
 from banzai.data import DataProduct, HeaderOnly, ArrayData, DataTable
 from banzai_floyds.orders import orders_from_fits
 from banzai_floyds.utils.wavelength_utils import WavelengthSolution
@@ -38,8 +37,8 @@ class FLOYDSObservationFrame(LCOObservationFrame):
         # Load the orders if they exist
         if 'ORDER_COEFFS' in self:
             self.orders = orders_from_fits(self['ORDER_COEFFS'].data, self['ORDER_COEFFS'].meta, self.shape)
-        if 'WAVELENGTHS' in self:
-            self.wavelengths = WavelengthSolution.from_header(self['WAVELENGTHS'].meta, self.orders)
+        if 'WAVELENGTH' in self:
+            self.wavelengths = WavelengthSolution.from_fits(self['WAVELENGTH'].meta, self.orders)
         if 'PROFILEFITS' in self:
             self.profile = load_profile_fits(self['PROFILEFITS'])
         if 'BINNED2D' in self:
@@ -219,7 +218,7 @@ class FLOYDSObservationFrame(LCOObservationFrame):
     @wavelengths.setter
     def wavelengths(self, value):
         self._wavelengths = value
-        self.add_or_update(HeaderOnly(value.to_header(), name='WAVELENGTHS'))
+        self.add_or_update(ArrayData(value.data, meta=value.to_header(), name='WAVELENGTH'))
 
     @property
     def elevation(self):