+++ banzai_floyds/wavelengths.py
@@ -6,15 +6,16 @@ from banzai_floyds.matched_filter import matched_filter_metric
 from scipy.signal import find_peaks
 from banzai_floyds.matched_filter import optimize_match_filter
 from banzai_floyds.frames import FLOYDSCalibrationFrame
-from banzai.data import ArrayData, DataTable
+from banzai.data import DataTable
 from banzai_floyds.utils.binning_utils import bin_data
 from banzai_floyds.utils.wavelength_utils import WavelengthSolution, tilt_coordinates
-from banzai_floyds.arc_lines import arc_lines_table, used_lines
+from banzai_floyds.utils.order_utils import get_order_2d_region
+from banzai_floyds.arc_lines import arc_lines_table
 from banzai_floyds.utils.fitting_utils import gauss, fwhm_to_sigma
 from banzai_floyds.extract import extract
 from scipy.special import erf
 from copy import copy
-from astropy.table import Table
+from astropy.table import Table, vstack
 from banzai.logs import get_logger
 
 
@@ -156,19 +157,19 @@ def refine_peak_centers(data, error, peaks, line_fwhm, domain=None):
 
 def correlate_peaks(peaks, linear_model, lines, match_threshold):
     """
-        Find the standard line peaks associated with the detected peaks in a raw 1D arc extraction
+    Find the standard line peaks associated with the detected peaks in a raw 1D arc extraction
 
-        Parameters
-        ----------
-        peaks: array containing the pixel location of detected peaks
-        linear_model: 1st order fit function for the wavelength solution
-        lines: table containing 'wavelength' and 'strength' for each standard line
-        match_threshold: maximum separation for a pair of peaks to be considered a match.
+    Parameters
+    ----------
+    peaks: array containing the pixel location of detected peaks
+    linear_model: 1st order fit function for the wavelength solution
+    lines: table containing 'wavelength' and 'strength' for each standard line
+    match_threshold: maximum separation for a pair of peaks to be considered a match.
 
-        Returns
-        -------
-        list of standard line peak wavelengths matching detected peaks
-        """
+    Returns
+    -------
+    list of standard line peak wavelengths matching detected peaks
+    """
     guessed_wavelengths = linear_model(peaks)
     corresponding_lines = []
     # correlate detected peaks to known wavelengths
@@ -192,122 +193,104 @@ def estimate_distortion(peaks, corresponding_wavelengths, domain, order=4):
     return Legendre.fit(deg=order, x=peaks, y=corresponding_wavelengths, domain=domain)
 
 
-def full_wavelength_solution_weights(theta, coordinates, lines, line_slices, bkg_order_x, bkg_order_y):
-    """
-    Produce a 2d model of arc fluxes given a line list and a wavelength solution polynomial, a tilt, and a line width
+def match_features(flux, flux_error, fwhm, wavelength_solution, min_line_separation, lines, match_threshold,
+                   domain=None, snr_threshold=5.0):
+    peaks = identify_peaks(flux, flux_error, fwhm, min_line_separation, domain=domain, snr_threshold=snr_threshold)
 
-    Parameters
-    ----------
-    theta: tuple: tilt, line_fwhm, *polynomial_coefficients, *background_coefficients, *line_strengths
-    coordinates: tuple of 2d arrays x, y. x and y are the coordinates of the data array for the model
-    lines: astropy table of the lines in the line list with wavelength (in angstroms) and strength
+    corresponding_lines = np.array(correlate_peaks(peaks, wavelength_solution, lines,
+                                                   match_threshold=match_threshold)).astype(float)
+    successful_matches = np.isfinite(corresponding_lines)
 
-    Returns
-    -------
-    model array: 2d array with the match filter weights given the wavelength solution model
-    """
-    tilt, line_fwhm = theta[:2]
-    n_bkg_coefficients = bkg_order_x + 1 + bkg_order_y + 1
-    polynomial_coefficients = theta[2:-n_bkg_coefficients - len(lines)]
-    bkg_x_low_index = 2 + len(polynomial_coefficients)
-    bkg_x_high_index = 2 + len(polynomial_coefficients) + 1 + bkg_order_x
-    bkg_coefficients_x = theta[bkg_x_low_index: bkg_x_high_index]
-    bkg_coefficients_y = theta[bkg_x_high_index:bkg_x_high_index + bkg_order_y + 1]
-    line_strengths = theta[-len(lines):]
-    x, y = coordinates
-    tilted_x = tilt_coordinates(tilt, x, y)
-    # We could cache the domain of the function
-    wavelength_polynomial = Legendre(polynomial_coefficients, domain=(np.min(x), np.max(x)))
-    model_wavelengths = wavelength_polynomial(tilted_x)
-    bkg_polynomial_x = Legendre(bkg_coefficients_x, domain=(np.min(x), np.max(x)))
-    bkg_polynomial_y = Legendre(bkg_coefficients_y, domain=(np.min(y), np.max(y)))
-
-    line_sigma = fwhm_to_sigma(line_fwhm)
-    # Convert line sigma in pixels to wavelengths
-    line_sigma = line_sigma * wavelength_polynomial.deriv(1)(tilted_x)
-    model = np.zeros(x.shape)
-    # Some possible optimizations are to truncate around each line (caching which indicies are for each line)
-    # say +- 5 sigma around each line
-    # We fit a relative strength of each line here to capture variations of the lamp
-    for line, line_slice, line_strength in zip(lines, line_slices, line_strengths):
-        # in principle we should set the resolution to be a constant, i.e. delta lambda / lambda, not the overall width
-        model[line_slice] += line_strength * gauss(model_wavelengths[line_slice], line['wavelength'],
-                                                   line_sigma[line_slice])
-
-    used_region = np.where(model != 0.0)
-    model[used_region] += bkg_polynomial_x(x[used_region]) * bkg_polynomial_y(y[used_region])
-    # TODO: There is probably some annoying normalization here that is unconstrained
-    # so we probably need 1 fewer free parameters
-    return model
+    peaks = refine_peak_centers(flux, flux_error, peaks[successful_matches], fwhm,
+                                domain=domain)
+    return peaks, corresponding_lines[successful_matches]
 
 
-def full_wavelength_solution(data, error, x, y, initial_polynomial_coefficients, initial_tilt, initial_line_fwhm,
-                             lines, background_order_x=4, background_order_y=2):
+def in_fitting_region(wavelengths, wavelength_derivs, lines, nsigma, sigma):
+    used_lines = []
+    line_regions = []
+    for line in lines:
+        wavelength_sigma = sigma * wavelength_derivs
+        in_line_region = wavelengths >= line['wavelength'] - nsigma * wavelength_sigma
+        in_line_region = np.logical_and(
+            in_line_region,
+            wavelengths <= line['wavelength'] + nsigma * wavelength_sigma
+        )
+        if in_line_region.sum() > 0:
+            used_lines.append(line)
+            line_regions.append(np.where(in_line_region))
+    return used_lines, line_regions
+
+
+def full_wavelength_solution(data, error, x, y, initial_polynomial_coefficients,
+                             initial_tilt_coefficients, line_fwhm,
+                             lines, domain, fitting_region_nsigma=5):
     """
     Use a match filter to estimate the best fit 2-d wavelength solution
 
     Parameters
     ----------
-    data: 2-d array with data to be fit
+    data: 2-d array with data to be fit ()
     error: 2-d array error, same shape as data
     x: 2-d array, x-coordinates of the data, same, shape as data
     y: 2-d array, y-coordinates of the data, same, shape as data
     initial_polynomial_coefficients: 1d array of the initial polynomial coefficients for the wavelength solution
-    initial_tilt: float: initial angle measured clockwise of up in degrees
-    initial_line_fwhm: float: initial estimate of fwhm of the lines in pixels
+    initial_tilt: 1d array: initial angle measured clockwise of up in degrees
+    line_fwhm: float: initial estimate of fwhm of the lines in pixels
     lines: astropy table: must have the columns of catalog center in angstroms, and strength
-    background_order_x: int: order of the polynomial in x to fit the background
-    background_order_y: int: order of the polynomial in y to fit the background
+    fitting_region: float: N-sigma away from the initital line position to include in fit
     Returns
     -------
-    best_fit_params: 1-d array: (best_fit_tilt, best_fit_line_width, *best_fit_polynomial_coefficients, *bkg_x, *bkg_y)
+    wavelength_polynomial, tilt_polynomial: Legednre polynomials
     """
-    # TODO: Add a backround component to the model so that the line widths are accurate
-    tilted_x = tilt_coordinates(initial_tilt, x, y)
-    # We could cache the domain of the function
-    wavelength_polynomial = Legendre(initial_polynomial_coefficients, domain=(np.min(x), np.max(x)))
-    line_sigma = fwhm_to_sigma(initial_line_fwhm)
-    pixels = np.arange(wavelength_polynomial.domain[0], wavelength_polynomial.domain[1] + 1, 1.0)
-    inverted_wavelength_polynomial = Legendre.fit(x=wavelength_polynomial(pixels), y=pixels,
-                                                  deg=wavelength_polynomial.degree(),
-                                                  domain=wavelength_polynomial.domain)
-    line_pixel_positions = np.array([inverted_wavelength_polynomial(line['wavelength']) for line in lines])
-    line_in_order = np.logical_and(np.min(x) <= line_pixel_positions, line_pixel_positions <= np.max(x))
-    lines = lines[line_in_order]
-    line_pixel_positions = line_pixel_positions[line_in_order]
-    # Cache where we think the lines are going to be. If we don't have a good initial solution this will create issues
-    # but that's true even if we don't cache here
-    line_slices = [np.where(np.logical_and(tilted_x > line_pixel - 5.0 * line_sigma,
-                                           tilted_x < line_pixel + 5.0 * line_sigma))
-                   for line_pixel in line_pixel_positions]
-    bkg_coefficients_x = np.zeros(background_order_x + 1)
-    bkg_coefficients_y = np.zeros(background_order_y + 1)
-    bkg_coefficients_x[0] = np.median(data)
-    bkg_coefficients_y[0] = 1.0
+    line_sigma = fwhm_to_sigma(line_fwhm)
+    initial_tilt_angle = Legendre(initial_tilt_coefficients, domain=domain)(x)
+    tilted_x = tilt_coordinates(initial_tilt_angle, x, y)
+    initial_wavelength_model = Legendre(initial_polynomial_coefficients, domain=domain)
+    initial_wavelengths = initial_wavelength_model(tilted_x)
+    initial_derivs = initial_wavelength_model.deriv()(tilted_x)
 
-    initial_line_strengths = []
+    used_lines, line_regions = in_fitting_region(
+        initial_wavelengths, initial_derivs, lines, fitting_region_nsigma, line_sigma
+    )
 
-    for line_pixel in line_pixel_positions:
-        line_flux = np.median(data[np.abs(tilted_x - line_pixel) <= 1.0])
-        line_flux -= bkg_coefficients_x[0]
-        line_flux /= (1 / np.sqrt(2.0 * np.pi) / line_sigma)
-        initial_line_strengths.append(line_flux)
+    to_fit = np.zeros_like(initial_wavelengths, dtype=bool)
+    for line_region in line_regions:
+        to_fit[line_region] = True
 
-    bounds = [(-90., 90), [0.0, None]]
-    for _ in initial_polynomial_coefficients:
-        bounds.append([None, None])
-    for _ in bkg_coefficients_x:
-        bounds.append([None, None])
-    for _ in bkg_coefficients_y:
-        bounds.append([None, None])
-    for _ in initial_line_strengths:
-        bounds.append([0.0, None])
-    best_fit_params = optimize_match_filter((initial_tilt, initial_line_fwhm, *initial_polynomial_coefficients,
-                                             *bkg_coefficients_x, *bkg_coefficients_y, *initial_line_strengths), data,
-                                            error, full_wavelength_solution_weights, (x, y),
-                                            args=(lines, line_slices, background_order_x, background_order_y),
-                                            bounds=bounds)
-    return best_fit_params
+    used_lines, line_regions = in_fitting_region(
+        initial_wavelengths[to_fit], initial_derivs[to_fit], used_lines, fitting_region_nsigma, line_sigma
+    )
+
+    fitted_data = data[to_fit] - np.median(data)
+    best_fit = optimize_match_filter(
+        [*initial_polynomial_coefficients, *initial_tilt_coefficients],
+        fitted_data, error[to_fit], wavelength_2d_weights, (x[to_fit], y[to_fit]),
+        args=(used_lines, domain, len(initial_polynomial_coefficients) - 1, line_regions, line_sigma)
+    )
+
+    wavelength_polynomial = Legendre(best_fit[:len(initial_polynomial_coefficients)],
+                                     domain=domain)
+    tilt_polynomial = Legendre(best_fit[len(initial_polynomial_coefficients):],
+                               domain=domain)
+    return wavelength_polynomial, tilt_polynomial
+
+
+def wavelength_2d_weights(theta, xy, lines, domain, wavelength_degree, line_regions, line_sigma):
+    x, y = xy
+    wavelength_coeffs = theta[:wavelength_degree + 1]
+    tilt_coeffs = theta[wavelength_degree + 1:]
+    tilt_polynomial = Legendre(tilt_coeffs, domain=domain)
+    wavelength_polynomial = Legendre(wavelength_coeffs, domain=domain)
+    weights = np.zeros_like(x, dtype=float)
+    for line, region in zip(lines, line_regions):
+        tilt_angle = tilt_polynomial(x[region])
+        tilted_x = tilt_coordinates(tilt_angle, x[region], y[region])
+        line_weights = gauss(wavelength_polynomial(tilted_x), line['wavelength'],
+                             line_sigma * wavelength_polynomial.deriv()(tilted_x))
+        line_weights *= line['strength']
+        weights[region] += line_weights
+    return weights
 
 
 class WavelengthSolutionLoader(FLOYDSCalibrationUser):
@@ -334,6 +317,7 @@ def estimate_line_centers(wavelengths, flux, flux_errors, lines, line_fwhm, line
     # Note line_separation is in pixels here.
     reference_wavelengths = []
     measured_wavelengths = []
+    pixel_positions = []
     peaks = np.array(identify_peaks(flux, flux_errors, line_fwhm, line_separation, snr_threshold=15.0))
     for line in lines:
         if line['wavelength'] > np.max(wavelengths) or line['wavelength'] < np.min(wavelengths):
@@ -346,56 +330,58 @@ def estimate_line_centers(wavelengths, flux, flux_errors, lines, line_fwhm, line
                 continue
             if np.abs(refined_peak - closest_peak) > 5:
                 continue
+            pixel_positions.append(refined_peak)
             refined_peak = np.interp(refined_peak, np.arange(len(wavelengths)), wavelengths)
             measured_wavelengths.append(refined_peak)
             reference_wavelengths.append(line['wavelength'])
-    return np.array(reference_wavelengths), np.array(measured_wavelengths)
+    return np.array(pixel_positions), np.array(reference_wavelengths), np.array(measured_wavelengths)
 
 
-def estimate_residuals(image, min_line_separation=5.0):
+def estimate_residuals(image, line_fwhm, used_lines, min_line_separation=10.0):
     # Note min_line_separation is in pixels here.
     reference_wavelengths = []
     measured_wavelengths = []
     orders = []
+    positions = []
 
     for order in [1, 2]:
         where_order = image.extracted['order'] == order
-        order_reference_wavelengths, order_measured_wavelengths = estimate_line_centers(
+        pixel_positions, order_reference_wavelengths, order_measured_wavelengths = estimate_line_centers(
             image.extracted['wavelength'][where_order],
             image.extracted['fluxraw'][where_order],
             image.extracted['fluxrawerr'][where_order],
-            used_lines, image.wavelengths.line_fwhms[order - 1],
+            used_lines, line_fwhm,
             min_line_separation
             )
         reference_wavelengths = np.hstack([reference_wavelengths, order_reference_wavelengths])
         measured_wavelengths = np.hstack([measured_wavelengths, order_measured_wavelengths])
+        positions = np.hstack([positions, pixel_positions])
         orders += [order] * len(order_reference_wavelengths)
-    return Table({'measured_wavelength': measured_wavelengths,
+    return Table({'position': positions,
+                  'measured_wavelength': measured_wavelengths,
                   'reference_wavelength': reference_wavelengths,
                   'order': orders})
 
 
 class CalibrateWavelengths(Stage):
     LINES = arc_lines_table()
-    # FWHM is in pixels
+    # FWHM is in pixels for the 2" slit
     INITIAL_LINE_FWHMS = {'coj': {1: 6.65, 2: 5.92}, 'ogg': {1: 4.78, 2: 5.02}}
     INITIAL_DISPERSIONS = {1: 3.51, 2: 1.72}
     # Tilts in degrees measured counterclockwise (right-handed coordinates)
     INITIAL_LINE_TILTS = {1: 8., 2: 8.}
-    OFFSET_RANGES = {1: np.arange(7200.0, 8000.0, 0.5), 2: np.arange(4300, 5200, 0.5)}
+    TILT_COEFF_ORDER = {'coj': 0, 'ogg': 0}
+    OFFSET_RANGES = {1: np.arange(7000.0, 7600.0, 0.5), 2: np.arange(4300, 5200, 0.5)}
     # These thresholds were set using the data processed by the characterization tests.
     # The notebook is in the diagnostics folder
     MATCH_THRESHOLDS = {1: 50.0, 2: 25.0}
     # In units of the line fwhm (converted to sigma)
-    MIN_LINE_SEPARATION_N_SIGMA = 7.5
+    MIN_LINE_SEPARATION_N_SIGMA = 5.0
     # In units of median signal to noise in the spectrum
     PEAK_SNR_THRESHOLD = 10.0
-    FIT_ORDERS = {1: 5, 2: 2}
+    FIT_ORDERS = {1: 4, 2: 2}
     # Success Metrics
     MATCH_SUCCESS_THRESHOLD = 3  # matched lines required to consider solution success
-    # Background polynomial orders
-    BKG_ORDER_X = 2
-    BKG_ORDER_Y = 2
     """
     Stage that uses Arcs to fit wavelength solution
     """
@@ -405,10 +391,8 @@ class CalibrateWavelengths(Stage):
         # Do a quick extraction by medianing the central region of the order
         extraction_orders = copy(image.orders)
 
-        best_fit_polynomials = []
-        best_fit_tilts = []
-        best_fit_fwhms = []
-
+        wavelength_polynomials = []
+        tilt_polynomials = []
         for i, order in enumerate(order_ids):
             order_region = extraction_orders.data == order
 
@@ -431,64 +415,52 @@ class CalibrateWavelengths(Stage):
                                          density=False)[0]
             flux_1d_error /= np.histogram(tilted_x, bins=bins, density=False)[0]
             flux_1d_error **= 0.5
+            # Convert to a 2" slit
+            initial_fwhm = self.INITIAL_LINE_FWHMS[image.site][order] * image.slit_width / 2.0
+
             linear_solution = linear_wavelength_solution(flux_1d, flux_1d_error, self.LINES[self.LINES['used']],
                                                          self.INITIAL_DISPERSIONS[order],
-                                                         self.INITIAL_LINE_FWHMS[image.site][order],
+                                                         initial_fwhm,
                                                          self.OFFSET_RANGES[order],
                                                          domain=image.orders.domains[i])
             # from 1D estimate linear solution
             # Estimate 1D distortion with higher order polynomials
-            min_line_separation = fwhm_to_sigma(self.INITIAL_LINE_FWHMS[image.site][order])
+            min_line_separation = fwhm_to_sigma(initial_fwhm)
             min_line_separation *= self.MIN_LINE_SEPARATION_N_SIGMA
-            peaks = identify_peaks(flux_1d, flux_1d_error,
-                                   self.INITIAL_LINE_FWHMS[image.site][order] / self.INITIAL_DISPERSIONS[order],
-                                   min_line_separation, domain=image.orders.domains[i],
-                                   snr_threshold=self.PEAK_SNR_THRESHOLD)
 
-            corresponding_lines = np.array(correlate_peaks(peaks, linear_solution, self.LINES[self.LINES['used']],
-                                                           self.MATCH_THRESHOLDS[order])).astype(float)
-            successful_matches = np.isfinite(corresponding_lines)
-            if successful_matches.size < self.MATCH_SUCCESS_THRESHOLD:
-                # TODO: Add Logging?
-                # too few lines for good wavelength solution
+            peaks, corresponding_lines = match_features(flux_1d, flux_1d_error, initial_fwhm, 
+                                                        linear_solution, min_line_separation,
+                                                        domain=image.orders.domains[i],
+                                                        lines=self.LINES[self.LINES['used']],
+                                                        match_threshold=self.MATCH_THRESHOLDS[order],
+                                                        snr_threshold=self.PEAK_SNR_THRESHOLD)
+            if len(peaks) < self.MATCH_SUCCESS_THRESHOLD:
+                logger.warning(f'Order {order} has too few matching lines for a good wavelength solution.')
                 image.is_bad = True
                 return image
 
-            peaks = refine_peak_centers(flux_1d, flux_1d_error, peaks[successful_matches],
-                                        self.INITIAL_LINE_FWHMS[image.site][order],
-                                        domain=image.orders.domains[i])
-
             initial_solution = estimate_distortion(peaks,
-                                                   corresponding_lines[successful_matches],
+                                                   corresponding_lines,
                                                    image.orders.domains[i],
                                                    order=self.FIT_ORDERS[order])
-            # Do a final fit that allows the fwhm, the line tilt, the strength of the catalog lines,
-            # the background, and a single set of polynomial coeffs,to vary.
-            # The background and line strengths are just nuisance parameters
-            # Limit the fit to only include +-5 sigma from known lines. This probably needs to be slit width
-            # dependent. This is so the lines that we don't include in the model don't pull the background fits.
-            # Fit 2D wavelength solution using initial guess either loaded or from 1D extraction
-            tilt, fwhm, *coefficients = full_wavelength_solution(
-                image.data[image.orders.data == order],
-                image.uncertainty[image.orders.data == order],
-                x2d[order_region], tilt_ys,
-                initial_solution.coef, self.INITIAL_LINE_TILTS[order],
-                self.INITIAL_LINE_FWHMS[image.site][order],
-                self.LINES[self.LINES['used']],
-                background_order_x=self.BKG_ORDER_X,
-                background_order_y=self.BKG_ORDER_Y
+            order_region_2d = get_order_2d_region(image.orders.data == order)
+            tilt_ys = y2d[order_region_2d] - image.orders.center(x2d[order_region_2d])[i]
+            # Do a final fit that allows the line tilt and a single set of polynomial coeffs to vary.
+            line_tilt_coeffs = np.zeros(self.TILT_COEFF_ORDER[image.site] + 1)
+            line_tilt_coeffs[0] = self.INITIAL_LINE_TILTS[order]
+            wavelength_polynomial, tilt_polynomial = full_wavelength_solution(
+                image.data[order_region_2d],
+                image.uncertainty[order_region_2d],
+                x2d[order_region_2d], tilt_ys,
+                initial_solution.coef,
+                line_tilt_coeffs,
+                initial_fwhm,
+                vstack([self.LINES[self.LINES['used']], self.LINES[self.LINES['blend']]]),
+                domain=image.orders.domains[i]
             )
-            # evaluate wavelength solution at all pixels in 2D order
-            # TODO: Make sure that the domain here doesn't mess up the tilts
-            polynomial = Legendre(coefficients[:self.FIT_ORDERS[order] + 1],
-                                  domain=(min(x2d[image.orders.data == order]),
-                                          max(x2d[image.orders.data == order])))
-            best_fit_polynomials.append(polynomial)
-            best_fit_tilts.append(tilt)
-            best_fit_fwhms.append(fwhm)
-        image.wavelengths = WavelengthSolution(best_fit_polynomials, best_fit_fwhms, best_fit_tilts, image.orders)
-        image.add_or_update(ArrayData(image.wavelengths.data, name='WAVELENGTHS',
-                                      meta=image.wavelengths.to_header()))
+            wavelength_polynomials.append(wavelength_polynomial)
+            tilt_polynomials.append(tilt_polynomial)
+        image.wavelengths = WavelengthSolution(wavelength_polynomials, tilt_polynomials, image.orders)
         image.is_master = True
 
         # Extract the data
@@ -498,8 +470,10 @@ class CalibrateWavelengths(Stage):
         binned_data['extraction_window'] = True
         image.extracted = extract(binned_data)
 
-        min_line_separation = fwhm_to_sigma(np.max(list(self.INITIAL_LINE_FWHMS[image.site].values())))
+        min_line_separation = fwhm_to_sigma(initial_fwhm)
         min_line_separation *= self.MIN_LINE_SEPARATION_N_SIGMA
-        image.add_or_update(DataTable(estimate_residuals(image, min_line_separation=min_line_separation),
+        image.add_or_update(DataTable(estimate_residuals(image, initial_fwhm,
+                                                         self.LINES[self.LINES['used']],
+                                                         min_line_separation=min_line_separation),
                                       name='LINESUSED'))
         return image